telegram-bot-deploy/requirements.txt
python-telegram-bot==20.0
httpx==0.24.0
nest-asyncio==1.5.4
